Added measurements minimization

+++ demo_linux/MainAgent.cpp
@@ -17,10 +17,9 @@
 // buffSize Number of element of retBuff.
 void libAgent(int retBuff[], size_t buffSize)
 {
-    // Dummy code
     // srand(time(NULL));
     runMeasure();
-    for (int i = 0; i < buffSize; i++)
+    for (size_t i = 0; i < buffSize; i++)
         // retBuff[i] = rand() % 10;
-        retBuff[i] = meas[i];
+        retBuff[i] = (int)(meas[i]);
 }
